alex-szarka/pyproject.toml
[project]
name = "alex-szarka"
version = "0.1.0"
description = "Student Certification Readiness Multi-Agent System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework",
    "azure-ai-agentserver-agentframework>=1.0.0b14",
    "azure-ai-projects>=2.0.0b1",
    "azure-identity",
    "python-dotenv",
]

[tool.uv]
prerelease = "allow"
